refactor: rename Groups to ListItems
