+++ src/popup/components/Views/MainView/ListItems/FolderItem.vue
@@ -2,8 +2,8 @@
 import type { Folder } from '@common/types'
 import { useRouter } from 'vue-router'
 import ChevronRightIcon from '@common/components/Icons/ChevronRightIcon.vue'
-import GroupIcon from '@/components/Views/MainView/Groups/GroupIcon.vue'
-import MainItem from '@/components/Views/MainView/Groups/MainItem.vue'
+import GroupIcon from '@/components/Views/MainView/ListItems/GroupIcon.vue'
+import MainItem from '@components/Views/MainView/ListItems/MainItem.vue'
 
 const props = defineProps<{ folder: Folder }>()
 const router = useRouter()
@@ -14,7 +14,10 @@ async function navigateToFolderView(): Promise<void> {
 </script>
 
 <template>
-    <MainItem @click="navigateToFolderView" class="bg-folder! hover:bg-folder-hover!">
+    <MainItem
+        @click="navigateToFolderView"
+        class="bg-folder! hover:bg-folder-hover!"
+    >
         <div class="flex items-center gap-2">
             <GroupIcon :folder />
             <h2 class="text-sm">{{ folder.name }}</h2>
