+++ src/popup/components/Views/MainView/ListItems/GroupIcon.vue
