+++ src/popup/components/Views/MainView/ListItems/ListItems.vue
@@ -2,10 +2,10 @@
 import type { Group, Folder } from '@common/types'
 import { trans } from '@common/modules'
 import { useGroupStore } from '@/stores/group'
-import GroupItem from '@/components/Views/MainView/Groups/GroupItem.vue'
+import GroupItem from '@/components/Views/MainView/ListItems/GroupItem.vue'
 import Message from '@common/components/Message.vue'
 import Spinner from '@common/components/Spinner.vue'
-import FolderItem from '@/components/Views/MainView/Groups/FolderItem.vue'
+import FolderItem from '@/components/Views/MainView/ListItems/FolderItem.vue'
 
 const groupStore = useGroupStore()
 
@@ -27,11 +27,7 @@ const props = withDefaults(defineProps<Props>(), {
     </Message>
 
     <div v-else class="flex flex-col">
-        <FolderItem
-            v-for="folder in folders"
-            :key="folder.name"
-            :folder
-        />
+        <FolderItem v-for="folder in folders" :key="folder.name" :folder />
 
         <div v-for="group in groups" :key="group.id">
             <GroupItem v-if="!group.hide" :group />
