+++ src/popup/components/Views/MainView/ListItems/OpenTabsButton.vue
